requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.2.0
colorama==0.4.6
fastapi==0.115.12
flake8==7.2.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
mccabe==0.7.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pycodestyle==2.13.0
pydantic==2.11.4
pydantic_core==2.33.2
pyflakes==3.3.2
PyJWT==2.10.1
pytest==8.3.5
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
